+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Change: Update links to coffee2code.com to be HTTPS
 * Unit tests:
     * New: Add test for default hooks
+    * New: Add tests for setting and query param names
     * Change: Remove unnecessary unregistering of hooks in `tearDown()`
     * Change: Move `test_turn_on_admin()` until just before first needed now that other tests can run before it
     * Change: Store plugin instance in class variable to simplify referencing it